+++ api/kroxylicious-filter-api/src/main/java/io/kroxylicious/proxy/filter/RequestFilter.java
@@ -12,7 +12,10 @@ import org.apache.kafka.common.protocol.ApiMessage;
 
 /**
  * A KrpcFilter implementation intended to simplify cases where we want to handle all or
- * most request types, for example to modify the request headers.
+ * most request types, for example to modify the request headers. If a Filter implements
+ * RequestFilter, it cannot also implement {@link CompositeFilter} or any of the specific
+ * message Filter interfaces like {@link ApiVersionsRequestFilter}. If a Filter implements
+ * RequestFilter, it may also implement {@link ResponseFilter}.
  */
 public interface RequestFilter extends KrpcFilter {